+++ web/src/components/Home/NovelImportDialog.tsx
@@ -138,7 +138,7 @@ export function NovelImportDialog({ open, novaDir, onOpenChange, onImported }: N
       <input
         ref={inputRef}
         type="file"
-        accept=".txt,.md,.markdown,text/plain,text/markdown,text/x-markdown"
+        accept=".txt,.md,.markdown,.chm,text/plain,text/markdown,text/x-markdown,application/vnd.ms-htmlhelp"
         className="hidden"
         onChange={(event) => void handleFileSelected(event.target.files?.[0])}
       />